rename integrationtests2 -> integrationtests
