+++ python-api/tests/integrationtests/reference_frames/COMPOSITE_DUAL_PREVIEW.png
