+++ python-api/tests/integrationtests/reference_frames/COMPOSITE_NONE.png
