+++ python-api/tests/integrationtests/reference_frames/COMPOSITE_PIP_PLACEMENT.png
