+++ python-api/tests/integrationtests/reference_frames/SPECTRASCOPE_110.png
