+++ python-api/tests/integrationtests/reference_frames/SWITCH_BLUE_GREEN.png
