+++ python-api/tests/integrationtests/reference_frames/SWITCH_GREEN_BLUE.png
