+++ python-api/tests/integrationtests/reference_frames/SWITCH_RED_GREEN.png
