--- python-api/tests/integrationtests2/__init__.py
@@ -1,3 +1,0 @@
-"""
-Integration tests for gst-switch and Python API
-"""
